added old method in case of issue in NFS

+++ ci_tools/client_setup.sh
@@ -41,16 +41,38 @@ function mount_robot_gui_tools {
 function install_chrome {
   # remove if any existing chrome is installed
   rpm -e google-chrome-stable
-
-  # install chrome locally
-  yes | sudo yum localinstall robot_chrome/google-chrome-stable_current_x86_64.rpm
+  if [[ -e "robot_chrome/google-chrome-stable_current_x86_64.rpm" ]]
+  then
+    echo '... google-chrome found on mounted dir, starting localinstall'
+    yes | sudo yum localinstall robot_chrome/google-chrome-stable_current_x86_64.rpm
+  else
+    echo '... google-chrome not found on mounted dir'
+    wget -N https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
+    check_installation
+  fi
   google-chrome --version
 }
 
 function install_chrome_driver {
   # copy chromedriver to bin & make it executable
-  yes | cp -rf  robot_chrome/chromedriver /usr/bin/
-  chmod +x /usr/bin/chromedriver
+  if [[ -e "robot_chrome/chromedriver" ]]
+  then
+    echo '... chromedriver found on mounted dir'
+    yes | cp -rf  robot_chrome/chromedriver /usr/bin/
+    chmod +x /usr/bin/chromedriver
+  else
+    echo '... chromedriver not found on mounted dir'
+    VERSION=$(curl http://chromedriver.storage.googleapis.com/LATEST_RELEASE)
+    wget -N https://chromedriver.storage.googleapis.com/$VERSION/chromedriver_linux64.zip
+    yes | unzip chromedriver_linux64.zip
+    yes | cp -rf  chromedriver /usr/bin/chromedriver
+    chmod +x /usr/bin/chromedriver
+  fi
+  if [[ -d "./virenv/bin" ]]
+  then
+    echo '... virenv/bin dir found'
+    yes | cp -rf  /usr/bin/chromedriver ./virenv/bin/
+  fi
   chromedriver --version
 }
